ENH: Support documents for better code formatting conformance
Use Update C++/C attributes for itk style conformance.

Add script to assist with running clang-format on topic-branches Derived
from
https://github.com/Kitware/CMake/blob/master/Utilities/Scripts/clang-format.bash

Define the .clang-format config file to enforce style that is compatible
with automated enforcement.

A code style that can be automatically enforced and integrated
with IDE's for automated conformance.

The code style options are defined at
https://clang.llvm.org/docs/ClangFormatStyleOptions.html